requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Werkzeug==3.0.1
Bootstrap-Flask==2.3.3
email-validator==2.1.0.post1
gunicorn==21.2.0
